feat: add promiseWithResolvers
